04/05

+++ app/Http/Controllers/Admin/PostController.php
@@ -9,63 +9,68 @@ use App\Models\Category;
 use App\Models\Post;
 use App\Models\User;
 use App\Repository\PostRepository;
+use App\traits\ImageTrait;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
-    protected $postRepository;
+    use ImageTrait;
 
-public function __construct()
+public function __construct(PostRepository $postRepository)
 {
-    $this->postRepository =  new PostRepository();
+    $this->postRepository =  $postRepository;
 }
-    // public function __construct(){
-    //     $this->middleware('permission: add posts| edit posts | delete posts| restore posts',['only'=>['posts']]);
-    //     $this->middleware('permission: add posts',['only'=>['addPosts','createPosts']]);
-    //     $this->middleware('permission: edit posts',['only'=>['editPosts','updatePosts']]);
-    //     $this->middleware('permission: delete posts',['only'=>['deletePosts']]);
-    //     $this->middleware('permission: restore post',['only'=>['trash','untrash']]);
 
-    // }
-    //
     public function posts(){
         $posts = Post::orderBy('id', 'desc')->with(['category','user'])->get();
-        // foreach ($posts as $post) {
-        //     $post->category->name = !is_null($post->category->name) ? $post->category->name : 'DG';
-        //     $post->user->firstName = !is_null($post->user->firstName) ? $post->user->firstName : 'no name';
-        // }
-        $data['users'] = User::get()->toArray();
-        $data['categories'] = Category::get()->toArray();
+        $data['users'] = User::get();
+        $data['categories'] = Category::get();
         return view('backend/posts/posts',$data,['posts' => $posts]);
     }
-    public function addPosts(){
+    public function addPosts(Request $request) {
+        // if (! Gate::allows('add_post')) {
+        //     abort(403);
+        // }
+        $this->authorize('add_post');
+
         $data['users'] = User::get()->toArray();
         $data['categories'] = Category::get()->toArray();
         return view('backend/posts/addposts',$data);
     }
     public function createPosts(PostRequest $postRequest){
-        $this->postRepository->createPost($postRequest);
+        $this->authorize('add_post');
+
+        $this->postRepository->createNewPost($postRequest);
         return redirect()->route('posts');
     }
     public function editPosts($id){
-        $post = Post::where('id', $id)->get()->toArray();
-        $users = User::all()->toArray();
+        $this->authorize('edit_post');
+        
+        $post = Post::where('id', $id)->get()->first();
         $categories = Category::all()->toArray();
-        return view('backend/posts/editposts',['post' => $post[0],'users'=>$users,'categories'=>$categories]);
+        return view('backend/posts/editposts',['post' => $post,'categories'=>$categories]);
     }
     public function updatePosts(EditPostRequest $editPostRequest, $id){
-
-        $this->postRepository->updatePost($editPostRequest, $id);
+        $this->authorize('edit_post');
+        // $post = Post::find($id);
+        $post = Post::whereId($id)->first();
+        $this->postRepository->updateNewPost($post,$editPostRequest);
         return redirect()->route('posts');
     }
 
     public function deletePosts(Request $request,$id) {
+        $this->authorize('delete_post');
+
         $post = Post::find($id);
         if(is_null($post)){
             $request->session()->flash('error', 'Bài viết không tồn tại');
             return redirect()->route('posts');
         }
         $post->delete();
+        return response()->json([
+            'code' => 200,
+        ]);
         // return redirect('admin/posts');
     }
 
@@ -84,16 +89,21 @@ public function __construct()
             ];
         Post::create($posts);
         return response()->json([
-            'status' => 200,
+            'code' => 200,
         ]);
     }
 
 
     public function trash(){
+        
+        $this->authorize('restore_post');
+
         $posts = Post::orderBy('deleted_at','desc')->with(['category','user'])->onlyTrashed()->get();
         return view('backend/posts/trash',['posts' => $posts]);
     }
     public function unTrash(Request $request, $id){
+        $this->authorize('restore_post');
+
         $post = Post::withTrashed()->find($id);
         $post->restore();
         $request->session()->flash('success', 'Đã khôi phục thành công!');
